Continue to improve how defaults work
This commit tries a different approach to set github-pages defaults.
Rather than doing it exclusively in code, which required merging the
github-pages hash and jekyll default configuration hash, we put most of
the defaults in a yaml file. The '_default_config.yml' file is packaged
with github-pages. When running the build command in bin, it
automatically adds the default config as the first config file, and then
adds the typical default file, (_config.yml in the source directory).
This commit also adds the ability to set custom config files in the
jekyll-v4-github-pages build command.

This should work more cleanly with jekyll, and provide a greater
ability to customize the jekyll configuration, both the defaults for
github-pages, and also the user's config.