Improve github-pages build command
This both simplifies the code used in the build command and improves it.
This is accomplished by reusing the code from within Jekyll that
processes which config files are used.

There are two big changes to this. The first is a bugfix. When including
a config file via cli argument, no source path is added onto it. This
behavior should not have been changed since this is not how it would
normally work via the cli. The second change is that config files with
'yaml' or 'toml' extensions will now be discovered as user default
config files if they are present. This again, is exactly how it works
normally with Jekyll.