Don't include _config.yml when it doesn't exist
The way this is intended to work is the same as GitHub's where you don't
need to include a _config.yml. Thats kind of the whole point so not sure
how I missed this. But this fixes this and makes sure any user defined
configs or the Jekyll default of _config.yml, exist before passing it
along.