Add jekyll-octicons dependency & bump rouge
Adds jekyll-octicons as a default dependency of pages-gem. It was being
added later on in the github pages chain of tools so it was available,
but just not from here in pages-gem. It seems the intent there *might*
have been to not pin octicons down to a specific version, therefore
making it easier to get the latest versions. This seems to have failed,
given it was pinned to 14.x while they are now up to 19.x. Not sure
where it was when I forked this, but still, definitely wasn't on the
latest major version.

Also bumped rouge to its latest version.